requirements.txt
streamlit==1.26.0
tensorflow==2.13.0
tensorflow-hub==0.13.0
Pillow==9.5.0
numpy==1.24.3
